drivers: mediatek: sensors-1.0: Fix pointer to integer cast
 ../../../../../../kernel/realme/mt6785/drivers/misc/mediatek/sensors-1.0/oplus_sensor_devinfo/sensor_devinfo.c:845:39:
 error: cast to smaller integer type 'uint32_t' (aka 'unsigned int') from
 'void *' [-Werror,-Wvoid-pointer-to-int-cast]
 DEVINFO_LOG("Ptr2UINT32(p) = %d \n", Ptr2UINT32(p));
                     ^~~~~~~~~~~~~

../../../../../../kernel/realme/mt6785/drivers/misc/mediatek/sensors-1.0/oplus_sensor_devinfo/sensor_devinfo.c:66:25: note: expanded from macro 'Ptr2UINT32'
                        ^~~~~~~~~~~~~

+++ drivers/misc/mediatek/sensors-1.0/oplus_sensor_devinfo/sensor_devinfo.c
@@ -51,7 +51,7 @@ enum {
 #define DEVINFO_LOG(fmt, args...)   pr_err(DEV_TAG"%s %d : "fmt, __func__, __LINE__, ##args)
 
 #define UINT2Ptr(n)     (uint32_t *)(n)
-#define Ptr2UINT32(p)   (uint32_t)(p)
+#define Ptr2UINT32(p)   (uintptr_t)(p)
 #define MAG_PARA_OFFSET               8
 /*#define SOURCE_NUM                    3*/
 #define MAG_PARA_NUM                  9
